# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Internship Management API",
        "description": "API for managing student internships and reports",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local server"
        },
        {
            "url": "https://yourdomain.com/api",
            "description": "Production server"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "description": "Authenticate user and return token",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User logout",
                "description": "Revoke current user's tokens",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/check-token": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Check token validity",
                "description": "Verify token and return user data",
                "operationId": "ff28c254629f841403194cc021a37182",
                "responses": {
                    "200": {
                        "description": "Token is valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/grades/set": {
            "post": {
                "tags": [
                    "Teacher"
                ],
                "summary": "Set grade for report",
                "description": "Set or update grade for student report",
                "operationId": "b697783e398c6ea0c0d19942be33f4a6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SetGradeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Grade set successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GradeResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reports/upload": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Upload student report",
                "description": "Upload or update internship report file",
                "operationId": "6bdf91ad031ea32cbf9e467b05a1146f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "internship_id",
                                    "file"
                                ],
                                "properties": {
                                    "internship_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Report updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Отчет успешно обновлен"
                                        },
                                        "report": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "student_id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "internship_id": {
                                                    "type": "integer",
                                                    "example": 13
                                                },
                                                "file_name": {
                                                    "type": "string",
                                                    "example": "report.pdf"
                                                },
                                                "file_path": {
                                                    "type": "string",
                                                    "example": "https://localhost/storage/report.pdf"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Report created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Отчет успешно загружен"
                                        },
                                        "report": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 63
                                                },
                                                "student_id": {
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "internship_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "file_name": {
                                                    "type": "string",
                                                    "example": "report.pdf"
                                                },
                                                "file_path": {
                                                    "type": "string",
                                                    "example": "https://localhost/storage/report.pdf"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "required": [
                    "login",
                    "password"
                ],
                "properties": {
                    "login": {
                        "type": "string",
                        "example": "teacher@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "LoginResponse": {
                "properties": {
                    "token": {
                        "type": "string",
                        "example": "1|abc123..."
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "login": {
                                "type": "string",
                                "example": "teacher@example.com"
                            },
                            "role": {
                                "type": "string",
                                "example": "teacher"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UserResponse": {
                "properties": {
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "login": {
                                "type": "string",
                                "example": "teacher@example.com"
                            },
                            "role": {
                                "type": "string",
                                "example": "teacher"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SetGradeRequest": {
                "required": [
                    "report_id",
                    "grade"
                ],
                "properties": {
                    "report_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "grade_type_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "comment": {
                        "type": "string",
                        "example": "Отличная работа!"
                    }
                },
                "type": "object"
            },
            "GradeResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Оценка успешно выставлена"
                    },
                    "grade": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "report_id": {
                                "type": "integer",
                                "example": 1
                            },
                            "grade": {
                                "type": "string",
                                "example": "Зачет"
                            },
                            "comment": {
                                "type": "string",
                                "example": "Отличная работа!"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Доступно только для студентов"
                    },
                    "message": {
                        "type": "string",
                        "example": "The provided credentials are incorrect."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}